proteomDjango/celery_Proteomics/requirements/requirements2.txt
django-admin-import
goatools
django-spaghetti-and-meatballs
celery-progress
django-markdownx
gunicorn
